Rename articleForm to PascalCase
