+++ app/javascript/article-form/components/MainImage.jsx
